Move mergeChecker to GitHub
It's used only by GitHub, makes more sense to be moved over
